Relocate globalStyled.js and change import path of globalStyled in App
