+++ src/components/globalStyled.js
